update text props to use long values

+++ packages/react-native/ReactCommon/react/renderer/attributedstring/conversions.h
@@ -1029,11 +1029,11 @@ inline MapBuffer toMapBuffer(const FontVariant& fontVariant) {
 inline MapBuffer toMapBuffer(const TextAttributes& textAttributes) {
   auto builder = MapBufferBuilder();
   if (textAttributes.foregroundColor) {
-    builder.putInt(
+    builder.putLong(
         TA_KEY_FOREGROUND_COLOR, toAndroidRepr(textAttributes.foregroundColor));
   }
   if (textAttributes.backgroundColor) {
-    builder.putInt(
+    builder.putLong(
         TA_KEY_BACKGROUND_COLOR, toAndroidRepr(textAttributes.backgroundColor));
   }
   if (!std::isnan(textAttributes.opacity)) {
@@ -1089,7 +1089,7 @@ inline MapBuffer toMapBuffer(const TextAttributes& textAttributes) {
 
   // Decoration
   if (textAttributes.textDecorationColor) {
-    builder.putInt(
+    builder.putLong(
         TA_KEY_TEXT_DECORATION_COLOR,
         toAndroidRepr(textAttributes.textDecorationColor));
   }
@@ -1110,7 +1110,7 @@ inline MapBuffer toMapBuffer(const TextAttributes& textAttributes) {
         TA_KEY_TEXT_SHADOW_RADIUS, textAttributes.textShadowRadius);
   }
   if (textAttributes.textShadowColor) {
-    builder.putInt(
+    builder.putLong(
         TA_KEY_TEXT_SHADOW_COLOR,
         toAndroidRepr(textAttributes.textShadowColor));
   }
